fixed experiment-navigator sample data on 32-bit machines
integer constant was pushed on stack as 32-bit integer, while
gtk_stree_store_set() expected a 64-bit integer.
this resulted in very strange crashes under 32-bit system

+++ lib/gtk-experiment-widgets/gtk-experiment-navigator.c
@@ -122,7 +122,7 @@ gtk_experiment_navigator_init(GtkExperimentNavigator *klass)
 	gtk_tree_store_append(store, &child, &toplevel);
 	gtk_tree_store_set(store, &child,
 			   COL_SECTION_NAME, "BAR",
-			   COL_TIME, 5*60*1000 /* 5 minutes */,
+			   COL_TIME, (gint64)5*60*1000 /* 5 minutes */,
 			   -1);
 
 	/*
